Only run the Powershell installation process if the software is not installed already

+++ bash-scripts/vm-post-deployment.sh
@@ -15,19 +15,23 @@ echo "########## Installing utility softwares ##########"
 sudo apt install -y nano curl wget nano git gnupg iputils-ping dnsutils bash-completion
 
 # ########## Install PowerShell ##########
-echo "########## Installing PowerShell ##########"
-# Update the list of packages
-sudo apt-get update
-# Install pre-requisite packages.
-sudo apt-get install -y wget apt-transport-https software-properties-common
-# Download the Microsoft repository GPG keys
-wget -q "https://packages.microsoft.com/config/ubuntu/$(lsb_release -rs)/packages-microsoft-prod.deb"
-# Register the Microsoft repository GPG keys
-sudo dpkg -i packages-microsoft-prod.deb
-# Update the list of packages after we added packages.microsoft.com
-sudo apt-get update
-# Install PowerShell
-sudo apt-get install -y powershell
+
+# Install PowerShell only if it is not already installed
+if [ ! -f /usr/bin/pwsh ]; then
+    echo "########## Installing PowerShell ##########"
+    # Update the list of packages
+    sudo apt-get update
+    # Install pre-requisite packages.
+    sudo apt-get install -y wget apt-transport-https software-properties-common
+    # Download the Microsoft repository GPG keys
+    wget -q "https://packages.microsoft.com/config/ubuntu/$(lsb_release -rs)/packages-microsoft-prod.deb"
+    # Register the Microsoft repository GPG keys
+    sudo dpkg -i packages-microsoft-prod.deb
+    # Update the list of packages after we added packages.microsoft.com
+    sudo apt-get update
+    # Install PowerShell
+    sudo apt-get install -y powershell
+fi
 
 # Restart the shell so that PATH changes take effect
 echo "########## System Update & Software Installation Done ##########"
